Draw the custom view and add it to the layout
